Probity/requirements.txt
asgiref==3.9.1
contourpy==1.3.3
cycler==0.12.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
fonttools==4.59.2
kiwisolver==1.4.9
matplotlib==3.10.5
mpmath==1.3.0
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tzdata==2025.2
